Initial stab at continuation API
This patch adds an initial implementation of a continuation API.  The
approach uses two basic structs — one for “blocks” and one for
“continuations”.  A block is a chunk of code that you can run; it can
produce a result, which is passed into a continuation.  We currently
have blocks for producing constant values, and a block that implements
an “if” statement.

There are some basic test cases of these blocks, but we could use some
more documentation.