Unify blocks and continuations
This patch removes the separate types for blocks and continuations; now
everything is a block.  The block execute function takes in an input
pointer, like the continuation function used to.  The execute function
no longer takes in a “dest” parameter; each block is now responsible for
keeping track of a “next” pointer, and the block type contains a new
set_next() method to update this pointer.  (We can't have a standardized
way to store the next pointers.  An if statement, for instance, doesn't
store the next pointer directly; instead, it has to update the next
pointer for its true and false branches.)