Initial implementation of “return” statement
This patch adds an initial implementation of a “return” statement.  This
starts a lot of the scaffolding that we'll need for actual function
calls and returns.  We now have a type for keeping track of the current
execution context.  Right now, this context only contains the return
address for the currently executing function, but eventually this will
also include a proper stack.  The return statement just transfers
control to the current return block.