More efficient stack representation
We now keep a pointer to the top of the stack, which makes it easier to
push values onto the stack.  In addition, the user of the stack is now
responsible for ensuring that there is enough space on the stack before
calling lgv_stack_push, reducing its cost.  Lastly, lgv_stack_push is
now defined as a macro, guaranteeing that it's always inlined.  These
modifications greatly reduce the size of the various block execution
functions that we have so far.