Pass stack top as additional execute parameter
The top-of-stack pointer is now passed along as a parameter to each
block's execute function.  This saves a couple of loads and stores.
We still maintain a top-of-stack pointer in the stack objects
themselves, but we only sync those back up in a halt statement.  We'll
probably also want to do a sync in each return statement, and each time
the stack is resized.