Block get_head method
This patch adds a new get_head method to the block interface.  This is
called when the block is used in a call to set_next (i.e., when it's the
“next” block).  This lets us skip over the no-op block objects at the
beginning of a control structure.  We need those no-ops to be able to
link everything together correctly, but their execute methods would
immediately just into the first element of the control structure.  With
this new method, the previous block can be linked directly into the
actual meat of the control structure.

As part of adding this method, I've also extracted out all of the method
pointers — except for execute — into a separate interface struct.  This
keeps the size of the instance structs small, but means that we need an
extra instruction to retrieve any method pointer in the iface struct.
This is why we keep execute's pointer in the instance struct; we need to
be able to access it as quickly as possible while executing the blocks.