Global state object
Many of the lagavulin methods took in a libcork garbage collector object
as their first parameter.  This has been moved inside a new global state
object, and we've made every function consistently take in a state
object as a first parameter.  Right now, the state object only contains
the garbage collector object, but it might grow in the future.

The state object has also been placed in the new libswanson library,
since it will be used both in the lagavulin implementation, and the
as-yet-unstarted LLVM implementation.