Scope objects
This patch starts the objects that will form the core of the Swanson0
language.  Swanson0 will be an explicitly-typed language, without a lot
of the syntactic sugar that Swanson itself will have.  (This also lets
me work on the low-level execution bits while the high-level language is
in flux.)

The first object is a _scope_ object, which provides an abstraction of a
symbol table.  Right now, a scope is a simple mapping of names (which
can be arbitrary strings) to other Swanson0 objects.