Macros, expressions, and execution engines
This patch adds three inter-related pieces to Swanson0.  Expressions are
chunks of executable code; they correspond to the blocks in the
lagavulin execution engine.  (In fact, the intention is that each
execution engine will subclass swan_expression to include any additional
details needed by the execution engine; lagavulin's expression object
includes a reference to a block.)  Macros take in arbitrary Swanson0
objects and generate expressions.  Finally, we have a class that defines
which functions any Swanson execution engine must provide.  Currently,
it must be able to execute an expression, and create a “kernel”.  The
kernel is a Swanson0 scope that provides access to the core macros that
implement the most basic features of the language.  Right now, there are
two simple macros; one for int constants, and one for the (+) operation
on ints.