Evaluating S0 code into a value
You can now evaluate an S0 AST tree to get the value that the code
produces.  All of the current instructions are supported.  This patch
also adds a new testing harness that lets us run executables multiple
times, each with a different stdin.  This is how we test the evaluation
code; we pass in S0 code files into stdin, and print out the resulting
type on stdout.  We then compare this to the expected results.