Single s0_type struct
We now have a single struct (s0_type), containing a union to hold all of
the details about each possible type.  This is more consistent with how
we implemented values and instructions.  This is a much simpler approach
than having separate structs for each kind of type, and since the set of
possible S0 types will be rather fixed, we don't need the ability to add
new S0 types at runtime.