MACRO values
This patch adds the MACRO instruction to the S0 language.  This
instruction creates a new macro value, which is a sequence of
instructions that can be evaluated later on an S0 file.  (There's no
instruction for doing this evaluation, though.)  Macros have upvalues,
parameters, and results.  The upvalues are provided when the macro is
created, and are shared across all evaluations of the macro.  The
parameters and results of a macro are types.  When the macro is
evaluated, we'll ensure that the actual parameters and results of the
macro satisfy these types.