CALL and RETURN instructions
This patch adds the CALL and RETURN instructions to the S0 language.
The CALL instruction evaluates the body of a macro.  The RETURN
instruction returns values from within the body of a macro.  The
top-level list of instructions in an S0 file now needs to have a RETURN
instruction at the end, explicitly stating which value is produced as
the result of evaluating the file.  (Before, the result was implicitly
the value produced by the last instruction.)