TCLASS instruction
This patch adds a new “class” type, along with a TCLASS instruction for
creating class types.  A class type is a mapping of names to values.
Classes can satisfy interface types if every entry in the interface
exists in the class, and the class's value for that entry satisfies the
interface's type for that entry.  (We don't have a function that checks
this property yet, though.)