Product types
This patch adds a new “product” type.  This is a simple tuple of values.
Functions and macros have been updated so that they no longer have a
list of parameters and results; instead, they have a single “input” type
and a single “output” type.  These can be tuples if you need to take in
more than one parameter or produce more than one result.  (The empty
product type now represents “void” for functions with no parameters or
results.)  Within the body of a macro, the single parameter is available
as “!0”.

There's no way to create instances of a product type yet, so macros
still take in a list of upvalues.