Macros have single upvalue
S0 macros can now only take in a single upvalue.  The upvalue is
optional; you can leave the UPVALUE clause out of the MACRO instruction
completely if you want to.  But if it's present, you can only give a
single upvalue.  To pass in multiple values, you can use a tuple value.