Objects and C functions
This patch introduces two new kinds of value.  An “object” is a
high-level value; it represents the things that S0 doesn't handle
directly on its own.  Each object is an instance of a class type; the
class type defines the operations that are valid on the object.

A “C function” is like a macro, in that it can be evaluated to produce
an S0 value, but instead of being implemented by a sequence of S0
instructions, its implemented by a custom C function.

This patch also adds the “prelude”, which is the collection of types and
values that are predefined for the Swanson language.  Right now, we've
defined a “bool” class type, “false” and “true” objects, and an “if” C
function.  The prelude is represented by a class type, since that's the
entity that we have that maps string keys to values.