Type satisfaction
This patch adds a new function for determining when one type “satisfies”
another.  If type α satisfies type β, we donote this as α <: β.  This
indicates that a value of type α can be used wherever a value of type β
is expected.  It's similar to the subtype relationship in an
object-oriented language.  For the most part, a type is only satisfied
by itself.  There are a few exceptions:

  • All types satisfy the ANY type

  • Any class or interface α can satisfy an interface β, assuming that
    all of the entries in β also exist in α, and that for each of these,
    the type of α's entry satisfies the type of β's entry.

  • A function type α satisfies another function type β if the output of
    α satisfies the output of β, and the input of β satisfies the input
    of α.  (The input types are contravariant.)