Initial Swanson metamodel
This patch adds support for a new metamodel for Swanson, largely based
on the Smalltalk metamodel.  Everything in Swanson is an object.
Objects are instances of classes.  Classes are instances of metaclasses.
In addition to the basic underlying C types for these entities, there
are a bunch of helper macros that will make it easier to define all of
the classes in the Swanson prelude.