Parsing metamodel method calls
This patch adds a new candidate S0 language which focuses on the new
metamodel.  Instead of a bunch of different low-level operations, you
can do exactly one thing in this version of S0: call a method.  The
results of the method call are stored in local variables, which can only
be accessed in the current block.  Variables are single-assignment; but,
the thing that a local variable refers to can be mutable.  (In other
words, each variable is somewhat similar to a "void * const" in C — you
can't update which thing is pointed to, but you can update that thing.)

We'll probably have to add two more capabilities to this language:
string literals and block literals.  But everything else should be
possible via method calls.